first tests

+++ src/main/java/com/example/apislaughterhouse/repository/AnimalRepository.java
@@ -3,5 +3,5 @@ package com.example.apislaughterhouse.repository;
 import com.example.apislaughterhouse.models.Animal;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AnimalRepository extends JpaRepository<Animal,Long> {
+public interface AnimalRepository extends JpaRepository<Animal, Integer> {
 }
